requirements.txt
ultralytics
torch
torchvision
opencv-python
numpy
pandas
shapely
orjson
lxml
tqdm
sumolib
traci
matplotlib
ultralytics
pyyaml